+++ apps/details-game-screen/vite.config.ts
@@ -1,12 +1,12 @@
 import { defineConfig, PluginOption } from 'vite';
 import solidPlugin from 'vite-plugin-solid';
-import devtools from 'solid-devtools/vite';
 
 export default defineConfig({
-  plugins: [devtools(), solidPlugin() as PluginOption],
+  plugins: [solidPlugin()] as PluginOption[],
   base: '/details-game',
   server: {
     port: 3004,
+    host: true,
   },
   build: {
     target: 'esnext',