+++ apps/games-list-screen/CHANGELOG.md
@@ -1,5 +1,18 @@
 # games-list-screen
 
+## 0.1.3
+
+### Patch Changes
+
+- b1efb10: chore(42): update eslint for all apps
+
+## 0.1.2
+
+### Patch Changes
+
+- 6b1b0fc: chore(39): add command check-types for all apps and fix problems
+- aaa25b7: chore(41): add create dockerfile for app game-list-screen
+
 ## 0.1.1
 
 ### Patch Changes