+++ apps/my-games-screen/CHANGELOG.md
@@ -1,5 +1,18 @@
 # my-games-screen
 
+## 0.1.3
+
+### Patch Changes
+
+- b1efb10: chore(42): update eslint for all apps
+
+## 0.1.2
+
+### Patch Changes
+
+- 6b1b0fc: chore(39): add command check-types for all apps and fix problems
+- ee2c214: chore(41): create Dockerfile from my-games-list-screen
+
 ## 0.1.1
 
 ### Patch Changes